+++ src/utils.ts
@@ -1,9 +1,9 @@
 export function color(elem: HTMLElement, value: string) {
   if (value.startsWith("--")) {
     const style = getComputedStyle(elem);
-    return style.getPropertyValue(value);
+    return style.getPropertyValue(value) || "";
   }
-  return value;
+  return value || "";
 }
 
 export function outerRect(rects: Rect[]) {