+++ src/listenable.ts
@@ -1,4 +1,4 @@
-type ListenableCallback<T> = (controller: Listenable<T>) => void;
+type ListenableCallback<T> = (context?: T) => void;
 
 export class Listenable<T> {
   listeners: ListenableCallback<T>[] = [];
@@ -11,9 +11,23 @@ export class Listenable<T> {
     this.listeners = this.listeners.filter((l) => l !== listener);
   }
 
-  notifyListeners() {
+  notifyListeners(value?: T) {
     for (const listener of this.listeners) {
-      listener(this);
+      listener(value);
     }
+  }
+}
+
+export class ValueNotifier<T> extends Listenable<T> {
+  value: T;
+
+  constructor(value: T) {
+    super();
+    this.value = value;
+  }
+
+  update(value: T) {
+    this.value = value;
+    this.notifyListeners();
   }
 }