+++ src/utils.ts
@@ -3,9 +3,9 @@ import { CanvasContext } from "./classes/widget";
 export function color(elem: HTMLElement, value: string) {
   if (value.startsWith("--")) {
     const style = getComputedStyle(elem);
-    return style.getPropertyValue(value);
+    return style.getPropertyValue(value) || "";
   }
-  return value;
+  return value || "";
 }
 
 export function outerRect(rects: Rect[]) {